[step7] 결제 API

+++ src/main/java/com/hanghae/concert/application/MemberService.java
@@ -15,7 +15,7 @@ public class MemberService {
     private final MemberQueryService memberQueryService;
     private final MemberCommandService memberCommandService;
 
-    public void updateMemberBalance(Long memberId, Long balance, PaymentType paymentType) {
+    public void updateMemberBalance(Long memberId, Integer balance, PaymentType paymentType) {
 
         // 유저 검증
         if (memberQueryService.existsMemberById(memberId)) {
